feature: added navigation to new page
