+++ components/skeleton/index.js
@@ -1,11 +1,13 @@
-import Head from "next/head";
-import Image from "next/image";
+import Link from "next/link";
 import styles from "./Skeleton.module.css";
+import { PAGES } from "/public/shared.constants";
 
 export default function Skeleton() {
   return (
-    <button className={styles.skeleton}>
+    <Link href={PAGES.NEW}>
+      <button className={styles.skeleton}>
         <h3 className="font-bold">Create your own</h3>
-    </button>
+      </button>
+    </Link>
   );
 }
