+++ public/shared.constants.js
@@ -1,0 +1,4 @@
+export const PAGES = {
+  HOME: "/",
+  NEW: "/new",
+};
